requirements.txt
pandas
numpy
erddapy
hvplot
dash
datashader
python-dotenv
PyYAML